Retrieve token from pin message payload
